commands/fsck: exit 1 on failed fsck
We currently exit successfully on a failed fsck, but we should instead
exit 1 so that users can tell that there's been a problem.  This also
means that users can easily use git lfs fsck in CI to check for invalid
pointers in their commits and have CI fail if there are.

Add some tests for our new functionality as well.

Reorder the existing code to reduce the need for an additional if
statement.

+++ commands/command_fsck.go
@@ -72,25 +72,24 @@ func fsckCommand(cmd *cobra.Command, args []string) {
 		return
 	}
 
-	if fsckDryRun {
-		return
+	if fsckDryRun || len(corruptOids) == 0 {
+		os.Exit(1)
 	}
 
-	if len(corruptOids) != 0 {
-		badDir := filepath.Join(cfg.LFSStorageDir(), "bad")
-		Print("objects: repair: moving corrupt objects to %s", badDir)
+	badDir := filepath.Join(cfg.LFSStorageDir(), "bad")
+	Print("objects: repair: moving corrupt objects to %s", badDir)
 
-		if err := tools.MkdirAll(badDir, cfg); err != nil {
+	if err := tools.MkdirAll(badDir, cfg); err != nil {
+		ExitWithError(err)
+	}
+
+	for _, oid := range corruptOids {
+		badFile := filepath.Join(badDir, oid)
+		if err := os.Rename(cfg.Filesystem().ObjectPathname(oid), badFile); err != nil {
 			ExitWithError(err)
 		}
-
-		for _, oid := range corruptOids {
-			badFile := filepath.Join(badDir, oid)
-			if err := os.Rename(cfg.Filesystem().ObjectPathname(oid), badFile); err != nil {
-				ExitWithError(err)
-			}
-		}
 	}
+	os.Exit(1)
 }
 
 // doFsckObjects checks that the objects in the given ref are correct and exist.
